+++ lib/pure/collections/lists.nim
@@ -458,15 +458,14 @@ proc prepend*[T](L: var SinglyLinkedList[T], value: T) {.inline.} =
     assert a.contains(9)
   prepend(L, newSinglyLinkedNode(value))
 
-proc concat*[T](L1: var SinglyLinkedList[T], L2: SinglyLinkedList[T]) {.since: (1, 5).} =
-  ## Concatenates (adds to the end) `L2` to `L1`. Efficiency: O(1).
+proc append*[T](L1: var SinglyLinkedList[T], L2: SinglyLinkedList[T]) {.since: (1, 5).} =
+  ## Appends (adds to the end) `L2` to `L1`. Efficiency: O(1).
   ## Note that the two lists share structure after the operation.
   runnableExamples:
-    import sequtils
     var a = [1, 2, 3].toSinglyLinkedList
     let b = [4, 5].toSinglyLinkedList
-    a.concat(b)
-    assert a.toSeq == [1, 2, 3, 4, 5]
+    a.append(b)
+    assert $a == "[1, 2, 3, 4, 5]"
   if L1.tail != nil:
     L1.tail.next = L2.head
   L1.tail = L2.tail
@@ -560,15 +559,14 @@ proc prepend*[T](L: var DoublyLinkedList[T], value: T) =
     assert a.contains(9)
   prepend(L, newDoublyLinkedNode(value))
 
-proc concat*[T](L1: var DoublyLinkedList[T], L2: DoublyLinkedList[T]) {.since: (1, 5).} =
-  ## Concatenates (adds to the end) `L2` to `L1`. Efficiency: O(1).
+proc append*[T](L1: var DoublyLinkedList[T], L2: DoublyLinkedList[T]) {.since: (1, 5).} =
+  ## Appends (adds to the end) `L2` to `L1`. Efficiency: O(1).
   ## Note that the two lists share structure after the operation.
   runnableExamples:
-    import sequtils
     var a = [1, 2, 3].toDoublyLinkedList
     let b = [4, 5].toDoublyLinkedList
-    a.concat(b)
-    assert a.toSeq == [1, 2, 3, 4, 5]
+    a.append(b)
+    assert $a == "[1, 2, 3, 4, 5]"
   if L1.tail != nil:
     L1.tail.next = L2.head
   L1.tail = L2.tail