+++ tests/stdlib/tlists.nim
@@ -94,16 +94,16 @@ block DoublyLinkedListConversion:
   doAssert $[1].toDoublyLinkedList == "[1]"
   doAssert $[1, 2, 3].toDoublyLinkedList == "[1, 2, 3]"
 
-block SinglyLinkedListConcat:
+block AppendingSinglyLinkedLists:
   block:
     var
       l0 = initSinglyLinkedList[int]()
       l1 = [1].toSinglyLinkedList
       l2 = [2, 3].toSinglyLinkedList
       l3 = [4, 5, 6].toSinglyLinkedList
-    l0.concat(l3)
-    l1.concat(l3)
-    l2.concat(l3)
+    l0.append(l3)
+    l1.append(l3)
+    l2.append(l3)
     doAssert $l0 == "[4, 5, 6]"
     doAssert $l1 == "[1, 4, 5, 6]"
     doAssert $l2 == "[2, 3, 4, 5, 6]"
@@ -113,21 +113,21 @@ block SinglyLinkedListConcat:
       l1 = [1].toSinglyLinkedList
       l2 = [2, 3].toSinglyLinkedList
       l3 = [4, 5, 6].toSinglyLinkedList
-    l3.concat(l0)
-    l2.concat(l1)
+    l3.append(l0)
+    l2.append(l1)
     doAssert $l3 == "[4, 5, 6]"
     doAssert $l2 == "[2, 3, 1]"
 
-block DoublyLinkedListConcat:
+block AppendingDoublyLinkedLists:
   block:
     var
       l0 = initDoublyLinkedList[int]()
       l1 = [1].toDoublyLinkedList
       l2 = [2, 3].toDoublyLinkedList
       l3 = [4, 5, 6].toDoublyLinkedList
-    l0.concat(l3)
-    l1.concat(l3)
-    l2.concat(l3)
+    l0.append(l3)
+    l1.append(l3)
+    l2.append(l3)
     doAssert $l0 == "[4, 5, 6]"
     doAssert $l1 == "[1, 4, 5, 6]"
     doAssert $l2 == "[2, 3, 4, 5, 6]"
@@ -137,7 +137,7 @@ block DoublyLinkedListConcat:
       l1 = [1].toDoublyLinkedList
       l2 = [2, 3].toDoublyLinkedList
       l3 = [4, 5, 6].toDoublyLinkedList
-    l3.concat(l0)
-    l2.concat(l1)
+    l3.append(l0)
+    l2.append(l1)
     doAssert $l3 == "[4, 5, 6]"
     doAssert $l2 == "[2, 3, 1]"